% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyAppUI.R
\docType{data}
\name{shinyAppUI}
\alias{shinyAppUI}
\title{shinyAppServer}
\format{
An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 3.
}
\usage{
shinyAppUI
}
\description{
shinyAppServer
}
\keyword{datasets}
