% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_gbif_dwc.R
\name{load_gbif_dwc}
\alias{load_gbif_dwc}
\title{Open GBIF download from zip file}
\usage{
load_gbif_dwc(zipfile = NA, tmpdir = NA, pgdriver = NA)
}
\arguments{
\item{zipfile}{Path to the zipfile}

\item{tmpdir}{Temporary path for extracted files}

\item{pgdriver}{ODBC driver for the Postgres database. If not set, the function will try some options.}
}
\description{
Function that loads the data in a zip file downloaded from GBIF
}
\examples{
\dontrun{
data_loaded <- load_gbif_dwc(zipfile = "####.zip", tmpdir = "/tmp")
}
}
